chore: divide testbench files
